Don't degenerate stacked plots for datasets without variation.
In such a case, the previous behavior let to a determination of
an empty data boundary, where the bottomLeft and topRight points
had equal x and/or y coordinates. This then lead to an invalid
transformation matrix in the coordinate plane, generating NaN
values when applying the transformation. Overall, this was
noticeably as an extreme lag when showing such data in a stacked
plot, which happens sometimes in heaptrack_gui.