CartesianAxis: Adapt ~all geometry when relevant  attributes change.
This is inefficient and unfortunate (and a bit ugly), but probably
required for correctness.

Fixes a bug where adding annotations to an axis with axis title
would not change the axis title's position to move away from the
then longer tick labels.

Afer fixing that, I applied the same fix to similar code.