tests: increate timeout of TestDrawIntoPainter to 600s
It can take a lot of time to process data_test.csv on very slow
architectures (e.g. mips*), possibly hitting the default QtTest timeout
(300 seconds).

Increate it to 600 seconds (10 minutes) in both QtTest and CTest (which
has a default timeout of 10 minutes, so we are setting it explicitly
now) to make sure also slower architectures are able to run this
successfully.