KGantt: Make KGantt adapt to palette changes to make it usable also with dark themes
Colors/brushes for text, grid, constraints etc is fetched from the app palette
instead of being hardcoded.
Task, event and summary task colors are not changed
as they work with both dark and light colors.

This makes it possible print on white paper even when using a dark style
as demonstrated with the GanttPalette example.

BUG:426240