Optimize: Add markers as rects to the mapper when possible
For scatter plots using the rect marker style, we don't need
to go through expensive QPainterPath just to build a polygon
for a simple rect.

When we plot lots of points then this can make a large performance
difference. That said, the reverse mapper is still going to eat
kill performance noticeably, just a bit less so...