Lazily populate the ReverseMapper scene
I.e. only build the expensive scene BSP when needed, which is only
rarely - i.e. when we request a tooltip. This way, repainting becomes
much faster as we only need to build the polygons, which is far faster
than building the full scene representation.

As an additional benefit, this approach allows us to easily merge
the polygons for the same indices, something that apparently
happens for stacked area charts. I.e. instead of multiple items for
the same index, we now only get a single one with a combined polygon.