Improve internationalization in KGantt
`QDate::toString()` does not fully localize dates:  it does not reorder
date components or choose the proper form of month names.  For example,
in a Polish locale it returns "maj 1" instead of "1 maja".
Furthermore, in Qt6 date and time `toString()` methods will not perform
localization.

This patch switches to `QLocale()::toString()`.

BUG: 453292